[Doris] Refactor tests
